requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
aiohttp>=3.8.0
websockets>=11.0.0
requests>=2.28.0
python-dotenv>=0.19.0
gunicorn==21.2.0
Werkzeug==2.3.7
